Outsourced 2FA Setup and added Modal
