+++ package.json
@@ -12,6 +12,7 @@
     "@fortawesome/fontawesome-svg-core": "^6.4.0",
     "@fortawesome/free-solid-svg-icons": "^6.4.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
+    "@headlessui/react": "^1.7.15",
     "@types/node": "20.2.1",
     "@types/react": "^18.2.6",
     "@types/react-dom": "18.2.4",
